OK. So this is the first version that works in an extremely minimal form. So, it displays one page.
But more importantly, it now does a huge amount behind the scenes. The code checks for a SQLite
pre-parsed version of the book. If it doesn't find it, it unpacks the epub, parses it, typesets
the content and writes out the typeset result to a sqlite database. It does this
asynchronously, allowing the main thread to get on with creating the rest of the interface,
which only blocks when the first page is requested.